requirements.txt
chromadb>=0.5.0
PyPDF2>=3.0.0
PyMuPDF>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
cohere>=5.0.0
streamlit>=1.30.0
httpx>=0.27.0